Move GraphicsPrimitives to separate module
